+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## TBD:
+* Return RT in CIAM public api #2854
+
 ## [2.8.1]
 * Update target device/OS when generating documentation.